+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/PushHandlerImpl.java
@@ -38,6 +38,11 @@ public class PushHandlerImpl implements PushHandler {
     }
 
     @Override
+    public void analysisReset() throws HandlerException {
+        pusher.analysisReset();
+    }
+
+    @Override
     public String toString() {
         return "PushHandlerImpl{" +
                 "pushers=" + pushers +