+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/ResetProcessor.java
@@ -1,7 +1,6 @@
 package com.dwarfeng.rbacds.impl.handler;
 
-import com.dwarfeng.rbacds.stack.handler.FilterLocalCacheHandler;
-import com.dwarfeng.rbacds.stack.handler.PushHandler;
+import com.dwarfeng.rbacds.stack.handler.*;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -19,14 +18,29 @@ public class ResetProcessor {
     private static final Logger LOGGER = LoggerFactory.getLogger(ResetProcessor.class);
 
     private final FilterLocalCacheHandler filterLocalCacheHandler;
+    private final ScopedRolePermissionAnalysisLocalCacheHandler scopedRolePermissionAnalysisLocalCacheHandler;
+    private final ScopedUserPermissionAnalysisLocalCacheHandler scopedUserPermissionAnalysisLocalCacheHandler;
+    private final ScopePermissionAnalysisLocalCacheHandler scopePermissionAnalysisLocalCacheHandler;
+    private final UserRoleAnalysisLocalCacheHandler userRoleAnalysisLocalCacheHandler;
+    private final PermissionUserAnalysisLocalCacheHandler permissionUserAnalysisLocalCacheHandler;
 
     private final PushHandler pushHandler;
 
     public ResetProcessor(
             FilterLocalCacheHandler filterLocalCacheHandler,
+            ScopedRolePermissionAnalysisLocalCacheHandler scopedRolePermissionAnalysisLocalCacheHandler,
+            ScopedUserPermissionAnalysisLocalCacheHandler scopedUserPermissionAnalysisLocalCacheHandler,
+            ScopePermissionAnalysisLocalCacheHandler scopePermissionAnalysisLocalCacheHandler,
+            UserRoleAnalysisLocalCacheHandler userRoleAnalysisLocalCacheHandler,
+            PermissionUserAnalysisLocalCacheHandler permissionUserAnalysisLocalCacheHandler,
             PushHandler pushHandler
     ) {
         this.filterLocalCacheHandler = filterLocalCacheHandler;
+        this.scopedRolePermissionAnalysisLocalCacheHandler = scopedRolePermissionAnalysisLocalCacheHandler;
+        this.scopedUserPermissionAnalysisLocalCacheHandler = scopedUserPermissionAnalysisLocalCacheHandler;
+        this.scopePermissionAnalysisLocalCacheHandler = scopePermissionAnalysisLocalCacheHandler;
+        this.userRoleAnalysisLocalCacheHandler = userRoleAnalysisLocalCacheHandler;
+        this.permissionUserAnalysisLocalCacheHandler = permissionUserAnalysisLocalCacheHandler;
         this.pushHandler = pushHandler;
     }
 
@@ -39,6 +53,22 @@ public class ResetProcessor {
             pushHandler.filterReset();
         } catch (Exception e) {
             LOGGER.warn("推送过滤功能重置消息时发生异常, 本次消息将不会被推送, 异常信息如下: ", e);
+        }
+    }
+
+    public void resetAnalysis() throws HandlerException {
+        // 重置分析结果。
+        scopedRolePermissionAnalysisLocalCacheHandler.clear();
+        scopedUserPermissionAnalysisLocalCacheHandler.clear();
+        scopePermissionAnalysisLocalCacheHandler.clear();
+        userRoleAnalysisLocalCacheHandler.clear();
+        permissionUserAnalysisLocalCacheHandler.clear();
+
+        // 推送消息。
+        try {
+            pushHandler.analysisReset();
+        } catch (Exception e) {
+            LOGGER.warn("推送分析结果重置消息时发生异常, 本次消息将不会被推送, 异常信息如下: ", e);
         }
     }
 }