+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/pusher/DrainPusher.java
@@ -23,6 +23,10 @@ public class DrainPusher extends AbstractPusher {
     }
 
     @Override
+    public void analysisReset() {
+    }
+
+    @Override
     public String toString() {
         return "DrainPusher{" +
                 "pusherType='" + pusherType + '\'' +