+++ rbac-distributed-service-impl/src/main/java/com/dwarfeng/rbacds/impl/handler/resetter/FixedDelayResetter.java
@@ -57,6 +57,7 @@ public class FixedDelayResetter extends AbstractResetter {
             try {
                 LOGGER.info("计划时间已到, 重置过滤功能...");
                 context.resetFilter();
+                context.resetAnalysis();
             } catch (Exception e) {
                 String message = "重置器 " + FixedDelayResetter.this +
                         " 执行重置调度时发生异常, 过滤功能将不会重置, 异常信息如下: ";