+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/handler/Pusher.java
@@ -24,4 +24,11 @@ public interface Pusher {
      * @since 2.0.0
      */
     void filterReset() throws HandlerException;
+
+    /**
+     * 分析结果重置时执行的推送操作。
+     *
+     * @since 2.0.0
+     */
+    void analysisReset() throws HandlerException;
 }