+++ rbac-distributed-service-sdk/src/main/java/com/dwarfeng/rbacds/sdk/handler/pusher/PusherAdapter.java
@@ -33,6 +33,11 @@ public abstract class PusherAdapter extends AbstractPusher {
     public void filterReset() throws HandlerException {
     }
 
+    @SuppressWarnings("RedundantThrows")
+    @Override
+    public void analysisReset() throws HandlerException {
+    }
+
     @Override
     public String toString() {
         return "PusherAdapter{" +